src/podHello/requirements.txt
fastapi
uuid
uvicorn
python-dotenv